app config

+++ app/src/main/java/io/github/bubinimara/davibet/data/DataRepositoryImpl.kt
@@ -1,6 +1,7 @@
 package io.github.bubinimara.davibet.data
 
 import android.util.Log
+import io.github.bubinimara.davibet.AppConfig
 import io.github.bubinimara.davibet.data.db.TweetDao
 import io.github.bubinimara.davibet.data.model.Tweet
 import io.github.bubinimara.davibet.data.network.ApiService
@@ -15,7 +16,7 @@ class DataRepositoryImpl(private val apiService: ApiService, private val dbServi
     }
 
     override fun getTweets(track: String):Flow<List<Tweet>>{
-        return flow  <List<Tweet>> {
+        return flow {
             dbService.removeTweets()
             // new coroutine in the same scope of the parent
             // if parent end this end too
@@ -31,14 +32,24 @@ class DataRepositoryImpl(private val apiService: ApiService, private val dbServi
                     Log.d(TAG, "INSERT: end insert")
                 }
 
-                launch (CoroutineName("Remove-Expired-Tweet")){
-                    while (true) {
-                        Log.d(TAG, "REMOVE: remove expired")
-                        dbService.removeExpired()
-                        delay(TweetDao.TWEET_LIFETIME)
+                if(AppConfig.DB_TWEET_LIFETIME>0) {
+                    launch(CoroutineName("Remove-Expired-Tweet")) {
+                        while (true) {
+                            Log.d(TAG, "REMOVE: remove expired")
+                            dbService.removeExpired()
+                            delay(AppConfig.DB_TWEET_LIFETIME)
+                        }
                     }
                 }
 
+                if(AppConfig.DB_TABLE_SIZE_CHECK_INTERVAL_TIME_MS>0) {
+                    launch(CoroutineName("Truncate-Table")) {
+                        while (true) {
+                            dbService.truncateTable()
+                            delay(AppConfig.DB_TABLE_SIZE_CHECK_INTERVAL_TIME_MS)
+                        }
+                    }
+                }
                 // block
                 try {
                     // emit all tweets from the database
@@ -50,10 +61,6 @@ class DataRepositoryImpl(private val apiService: ApiService, private val dbServi
                 // todo:free all resources
                 Log.d(TAG, "getTweets: ALL END")
             }
-        }.catch { e->
-            Log.e(TAG, "getTweets: ",e )
-            // handle error
-            throw e
         }.flowOn(Dispatchers.IO)
     }
 }
